Refactor sso folder to extensions
